Football: Scoring After Game ended Goal will not be counted
Scoring after the game has ended will not be counted as a new goal. The ball has been shooted before the game ended, but is still on its way to the goal. The ceeper can not move anymore and this should not be counted as goal - as the player would be able to lose the game (if there is only 1 live left)